fix: fixing isVet and adding tests to guarantee this
