+++ app/views/pet/all.phtml
@@ -17,7 +17,7 @@
             <h1 class="font-bold text-2xl"><?= $name ?> Pets</h1>
             <div class="grid md:grid-cols-2 lg:grid-cols-3 gap-4">
                 <?php foreach ($pets as $pet): ?>
-                    <a href="<?= $loggedUserId == $userId ? route("pet.update", ["id" => $pet->id]) : "" ?>" class="card p-4 flex flex-col hover:bg-blue-500/10">
+                    <a href="<?= $loggedUserId == $userId ? route("pet.update", ["id" => $pet->id]) : "" ?>" class="<?= $loggedUserId == $userId ? "" : "pointer-events-none" ?> card p-4 flex flex-col hover:bg-blue-500/15 transition-all duration-300">
                         <h2 class="font-bold"><?= $pet->name ?></h2>
                         <?php if ($loggedUserId == $userId): ?>
                             <p>Edit or Delete</p>
@@ -25,7 +25,7 @@
                     </a>
                 <?php endforeach; ?>
                 <?php if ($loggedUserId == $userId): ?>
-                    <a href="<?= route("pet.create") ?>" class="card p-4 flex flex-col justify-center hover:bg-blue-500/10 transition-all duration-300">
+                    <a href="<?= route("pet.create") ?>" class="card p-4 flex flex-col justify-center hover:bg-blue-500/15 transition-all duration-300">
                         <h2 class="font-bold">New Pet</h2>
                     </a>
                 <?php endif; ?>
